Sentry should report errors from the client
Fixes https://github.com/lithictech/webhookdb-api/issues/289

If we get a 404, the server doesn't error,
but the CLI certainly should.

This adds a Resty middleware that will call a handler for 404s and 500s
which will configure and call Sentry.
See https://sentry.io/organizations/lithic-technology/issues/3189363997
for an example.