Do not run auth state machine if complete
Fixes https://github.com/lithictech/webhookdb-api/issues/428

Auth is a complex state machine flow,
where it can run its own state machine,
or implicit one due to the error-response-code.

If it is the latter, and the step is complete,
then do not try to re-run the step, since it results
in a duplicate message being printed.