Fix duplication in some step output
Fixes https://github.com/lithictech/webhookdb-api/issues/616

I'm not sure why, but when a state machine was run,
if the step was complete,
we'd always print output and finish the step.

This meant that if you pass in a completed step,
you could get the output printed again,
because the output *also* prints when the step is returned
from the API.

Instead, if we try to transition a completed step,
we should no-op, rather than print.

Additionally, one bug with the CLI is that invalid fields
(for example the `role` in `wh org invite --role=foo`)
would 400, and the state machine would fail the transition
and re-prompt for the org member email (because POSTing the email
failed the transition, but not because the email was invalid).

Instead, with https://github.com/lithictech/webhookdb-api/pull/652
we have a better validation error shape. This allows us to check that
the parameter being POSTed for the transition is one of the parameters
that failed validation. If it is, then we're safe to re-prompt.
If it's NOT a parameter that failed validation,
we know the caller can't fix the problem from the state machine,
it requires a new CLI call, so we can exit.